src/backend/requirements.txt
Flask==3.0.2
numpy==1.26.3
pandas==2.2.0
pyarrow==15.0.0
scikit-learn==1.4.0
python-dotenv==1.0.1